+++ cpsign-api/src/main/java/com/arosbio/cheminf/filter/HACFilter.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (C) Aros Bio AB.
+ *
+ * CPSign is an Open Source Software that is dual licensed to allow you to choose a license that best suits your requirements:
+ *
+ * 1) GPLv3 (GNU General Public License Version 3) with Additional Terms, including an attribution clause as well as a limitation to use the software for commercial purposes.
+ *
+ * 2) CPSign Proprietary License that allows you to use CPSign for commercial activities, such as in a revenue-generating operation or environment, or integrate CPSign in your proprietary software without worrying about disclosing the source code of your proprietary software, which is required if you choose to use the software under GPLv3 license. See arosbio.com/cpsign/commercial-license for details.
+ */
 package com.arosbio.cheminf.filter;
 
 import java.util.ArrayList;
@@ -36,6 +45,10 @@ public class HACFilter extends BaseFilter<HACFilter> implements Aliased {
         return NAME;
     }
 
+    public String toString(){
+        return "HAC filter: ["+minHAC+".."+maxHAC+']';
+    }
+
     @Override
     public String[] getAliases() {
         return new String[]{ALIAS};
@@ -45,10 +58,17 @@ public class HACFilter extends BaseFilter<HACFilter> implements Aliased {
         this.minHAC = min;
         return this;
     }
+    public int getMinHAC(){
+        return minHAC != null? minHAC : 0;
+    }
 
     public HACFilter withMaxHAC(Integer max){
         this.maxHAC = max;
         return this;
+    }
+
+    public int getMaxHAC(){
+        return maxHAC != null? maxHAC : Integer.MAX_VALUE;
     }
 
     @Override
@@ -110,6 +130,14 @@ public class HACFilter extends BaseFilter<HACFilter> implements Aliased {
         }
         
         throw new IllegalArgumentException("Molecule should not be discarded");
+    }
+
+    @Override
+    public HACFilter clone(){
+        return new HACFilter()
+            .withApplyToPredictions(applyToPredictions())
+            .withMaxHAC(maxHAC)
+            .withMinHAC(minHAC);
     }
     
 }