+++ cpsign-api/src/main/java/com/arosbio/cheminf/filter/MolMassFilter.java
@@ -1,3 +1,12 @@
+/*
+ * Copyright (C) Aros Bio AB.
+ *
+ * CPSign is an Open Source Software that is dual licensed to allow you to choose a license that best suits your requirements:
+ *
+ * 1) GPLv3 (GNU General Public License Version 3) with Additional Terms, including an attribution clause as well as a limitation to use the software for commercial purposes.
+ *
+ * 2) CPSign Proprietary License that allows you to use CPSign for commercial activities, such as in a revenue-generating operation or environment, or integrate CPSign in your proprietary software without worrying about disclosing the source code of your proprietary software, which is required if you choose to use the software under GPLv3 license. See arosbio.com/cpsign/commercial-license for details.
+ */
 package com.arosbio.cheminf.filter;
 
 import java.util.ArrayList;
@@ -34,6 +43,10 @@ public class MolMassFilter extends BaseFilter<MolMassFilter> implements Aliased 
     @Override
     public String getName() {
         return NAME;
+    }
+
+     public String toString(){
+        return "MolMass filter: ["+minMass+".."+maxMass+']';
     }
 
     @Override
@@ -112,6 +125,12 @@ public class MolMassFilter extends BaseFilter<MolMassFilter> implements Aliased 
     }
 
 
-
+    @Override
+    public MolMassFilter clone(){
+        return new MolMassFilter()
+            .withApplyToPredictions(applyToPredictions())
+            .withMinMass(minMass)
+            .withMaxMass(maxMass);
+    }
     
 }